Created new component called motorcycles to have container component be only a container that holds other components
