+++ src/components/RemoveFilterButton.js
@@ -3,11 +3,11 @@ const RemoveFilterButton = ({ getMotoData }) => {
         return () => {
             console.log('remove filter clicked!');
             getMotoData()
-        }
-    }
+        };
+    };
     return (
         <li id='remove-filter' onClick={removeFilterClicked()}>remove filter</li>
-    )
+    );
 };
 
 export default RemoveFilterButton;
